Update README.md
The Personal Portfolio Website is a responsive and interactive web application developed using HTML, CSS, and JavaScript. It serves as an online representation of an individual’s professional profile, showcasing their skills, projects, experience, and contact details. The website is designed to create a strong first impression for recruiters, clients, or collaborators.